fix(Combobox): fix normalize functionality for combobox
Recently we introduced the functionality to normalize the search results of the combobox. This means that the search value will always fall back to regular and lowercase characters instead of special characters. There was only one problem: using special characters wouldn't match the search results. This commit wil fix that.