fix(button): fix the text-shadow on the button component
After #1077 was merged we noticed some text-shadow issues on the button component in dark mode.
Turned out the shadow.text property was using a variable that was changed in #1077 causing a weird
text-shadow in dark mode. This commit will fix it by using a dedicated shadow for the button, after
discussing with Design, and only using it for the primary variant