feat(cover): expose the Content wrapper as own component
We have some pre-defined styles for the content inside the cover, mostly spacing. Sometimes you want
to alter these styles but that's only possible by styling the cover and using child selectors. This
commit exposes the Cover Content component so it's possible to style this specific component.

BREAKING CHANGE: The Content component, that's handling the spacing of the content of the cover, is
now exposes as it's own component and not added by default anymore.