fix: add pre-semantic-release to copy package.json to dist folder before running semantic-release
Semantic release requires there to be a package.json file in the package root folder (dist). Our
build process is copying this package.json file but this is done after semantic release does it's
validation steps. Therefor we need to add the package.json to the dist folder BEFORE running the
actual build steps.