fix(DatePicker): Calendar not updated when date prop changed indirectly
This commit fixes an issue where the calendar was not being
displayed correctly when the given date changed indirectly (i.e. not by
clicking and selecting the date before, but by passing a new prop eventually
changed by another component). This commit also introduces a specific test
for this scenario.

JIRA GIN-1718