refactor: Refactor useTransition hook and fix memory leaks
This commit re-factors the useTransition hook. Important changes made:
- The transitioning state is removed since we can use the "state" state
to determine that;
- rAF inside a rAF is not necessary if we useLayoutEffect. Contrary to
useEffect, which fires after the browser has painted, useLayoutEffect
fires before the painting phase of the browser. One rAF call is still
necessary to schedule the layout update for the current frame.
https://reactjs.org/docs/hooks-reference.html#uselayouteffect
https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame
https://blog.jakuba.net/request-animation-frame-and-use-effect-vs-use-layout-effect/
- useCallback does not have any cleanup functions like useEffect, contrary
to what we had in the code. Instead we now clear the timers and scheduled
rAF when the component unmounts (needs an empty dependency array to
fire only once and on unmount).