feat: introduce new semantic colors system
This feature will remove our current colors system of how we know it and replace it with a new, more
semantic colors system. This should make it easier to be consistent across our product and will make
it also easier to work with dark mode.

BREAKING CHANGE: Almost all our colors will be removed, so you need to replace it with other colors.
A list of removed colors can be found in the PR here: https://github.com/TicketSwap/solar/pull/1345.
The new semantic colors system can be found here:
https://ticketswap.github.io/solar/?path=/story/guidelines-colors--page.