fix(Card): pass verticalAlign prop to Body
The verticalAlign prop was being used by the body component in the CSS
but was not passed to the component so it was always centring the
content. This commit makes sure that the alignment is passed to the
body component and will make sure that it's possible again to
vertical align the cards.