fix(Dialog): lower the z-index to show toasts
We always want to show the toasts on top of the dialog. Because they
have the same z-index currently, it really depends on the html structure
which element is shown on top. By making sure the dialog overlay has
a lower z-index than the toast, the toast will always be shown on top
of the dialog.