feat(Toast): Make sure to use inverted colors for toast
Design would like to use inverted colors for our toasts, so that they
are better visible. This was already done before, and reverted, but
this was probably done because the colors weren't correctly displayed
in storybook. This is because the toasts are displayed in a portal,
and storybook is not handling them really great. Added #1608 to improve
this.