fix(Accordion): Extend Accordion props with Reach props
Not sure why but we were using our own type for the Accordion props. I
thought it was a good idea to use the props from Reach itself, but found
out that the `onChange` prop has an optional `index`. Created a PR
for the reach package to fix that, but for now we can extend the Reach
props and type the onChange prop ourself. This will benefit us, because
then we can make the accordion controlled.