fix(chipwithselect): Introduced name prop to ChipWithSelect options
In theory we would have a label everywhere, but in action (in sirius) the label is actually a `name` props as it is name of the period or the city or the category. While it could be updated in the category and period filter in the frontend, the backend is returning the list of location with city name instead of city labels. This way we support both.