fix(Toast): fix re-render issues of notify
There were a few issues with the notify method of the toast provider,
one of them was that the notify method was not being memoized correctly
and the other problem was that we didn't set the state correctly when
adding a new item, aka when notifying. This fixes the problem.